Use a symlink cache for the flatpak AppStream support
The idea is to encode the scope and user-specified origin in the filename as
flatpak remotes have metadata with the filename `appstream.xml.gz` that have
origin harcoded to 'flatpak'. This requires the corresponding appstream-glib
functionality to work.

The gnome-software process will keep the symlink cache up to date. I'm using
the AppStream-specified ~/.local/share/app-info/xmls at the moment, but if we
decide that the data is actually gnome-software specific then it can move.